// autogpt_platform/frontend/src/app/api/__generated__/models/blockCost.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { BlockCostCostFilter } from "./blockCostCostFilter";
import type { BlockCostType } from "./blockCostType";

export interface BlockCost {
  cost_amount: number;
  cost_filter: BlockCostCostFilter;
  cost_type: BlockCostType;
}