// autogpt_platform/frontend/src/app/api/__generated__/models/blockCostType.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */

export type BlockCostType = (typeof BlockCostType)[keyof typeof BlockCostType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const BlockCostType = {
  run: "run",
  byte: "byte",
  second: "second",
} as const;